vim(mappings): remap `+` nmap to always turn on signcolumn
`signcolumn` is a local window option, so if we open the same buffer on
different window, `signcolumn` might not be turned on. So add normal
mode mapping to always turn on `signcolumn` on the buffer in different
window.

As for default normal mode `+` keybinding, it has the same behavior as
`enter` normal mode keybinding so it's basically useless for me.