vim(lsp): change go to definition nmap to `gb`
To prevent conflict with default normal mode keybinding go to
definition of tag, change the vim-lsp go to definition to `gb`.

`gb` normal mode keybinding is used in netrw instead of default vim
keybinding, so there should be no problem. And as far as i know, vim-lsp
won't interfere with netrw.