shell(gitrc): use if-else in ggl function when no remote branch
The problem is that, when we have unstaged changes and we use `ggl`, git
will throw an error about cannot pull because of unstaged changes. And
because of that, the previous logic will unset the upstream branch.

So use if-else instead of logical expression to make the logic easier to
read and manage.