tmux(conf): prevent create multiple window using ctrl-w keybinding
`ctrl-w` is to spawn my preferred layout, which is 2 window with 1 pane
on first window and 2 pane on second window. Add target window when
create new window with `ctrl-w` keybinding so that when i accidentally
press `prefix ctrl-w`, it won't create new window and throw an error
instead.

Also trying to use line continuation to split out the command.