tmux(conf): use select-layout for ctrl-w keybinding
Percentage with `split-window -l` is a new feature in version 3.1a,
previously when we want to use percentage, we use `-p`. So if we use
`ctrl-w` keybinding on older tmux version, it will throw an error.

To prevent that, use `select-layout even-horizontal` to spread out the
2 panes evenly. Please keep in mind that `select-layout even-horizontal`
is not exactly the same as `split-window -h -l 50%`. I can replace
`split-window -h -l 50%` with `select-layout even-horizontal` because
there are only 2 panes in one window.