vim(mappings): change checktime nmap to `ZA`
Just remembered that default normal mode underscore mapping is useful,
because it's has almost the same behavior as `^`, which is go to the
first non-blank character on line, the difference is that underscore
mapping can use count to make it go downward like enter normal mode
mapping.

So, remap checktime to `ZA` instead.