vim(projections): add example of related property
Just found out that there's a `related` property that let us use the
navigation command without any argument.

In this example, when we are in the file
`after/autoload/possession.vim` and there's a file
`plugin/possession.vim`, we can use `Eplugin` (without argument) to open
the `plugin/possession.vim` directly.

Please keep in mind if the file does not exist, it will open an empty
buffer.