vim(defaults): prevent detect indent on csv and large files
Just found out that using `search()` in 1 GB csv file is slow. Not sure
where the limit is, so let's assume any file more than or equal 100 MB
is large.

With that in mind, prevent running detect indent, which uses `search()`,
in csv and large files.

I don't think that i would need precise indentation in a csv file, so it
should be no problem. The only thing i need to watch out is when editing
large files, which i haven't done except editing csv files.