vim(defaults): use default grepformat
Just realized that `grepformat` option is only effective when we use
`grep` command. I use `cgetexpr` command instead of `grep` command to
make `grep` (external tools, not built-in vim command) or `ripgrep` not
spit out the result in terminal, basically make grepping in the
background.

Because it's redundant, remove `grepformat` option from config and use
the default value of `grepformat`.