vim(mappings): use getloclist and getqflist for toggle window
The problem with `getwininfo()` approach is that we can't distinguish
between location list window or quickfix window when filtering the
quickfix window.

For example, let's say we have location list window open, and we press
the toggle for quickfix window. What happen is that the toggle for
quickfix window assume we have quickfix window open, even though the one
being opened is location list window.

I know that i should use more distinguishable keybindings, but
nevertheless the behavior when using `getwininfo()` is quite confusing.

So, let's try using `getloclist()` and `getqflist()` to toggle for
location list and quickfix window.