vim(defaults): prevent error when shiftwidth = 0 when set tabstop
When set tabstop with detect indent, there's an edge where default vim
shiftwidth for specific filetype is set to 0. When that happen, there's
an error `E487`, which basically means we need a positive number for
tabstop value.

So, to prevent that error, use global shiftwidth value when set tabstop
with detect indent.