vim(mappings): use `+` nmap to re-sync syntax highlight
Because vim use regex to do syntax highlight, so if the file is large,
sometimes the syntax highlight is messed up.

One of the solution is to re-sync the syntax highlight with command
`:syntax sync ...`.

Also, i replace the previous `+` normal mode keybinding because i feel
like i don't really need to set signcolumn all the time.