zsh(zshrc): replace inc_append_history with `fc -AI` in `preexec`
The problem with `inc_append_history` option is that it's append the
command that already in history file multiple times until we exit the
shell (maybe triggering `zshexit` hooks?). That behavior will increase
the history file size quickly if we use multiple zsh session.

So let's try using `preexec` hook and `fc -AI` to replace
`inc_append_history` and see how it goes.

Also try decreasing `HISTSIZE` and `SAVEHIST` to 100000 and see if it's
enough.