shell(aliasrc): open vim in `sc` alias with fzf `become()`
Just found out that fzf has `become()` which act like `execute()` but
will not come back to fzf after the execution. Basically it will turn
fzf into a new process of the command.

We can still use `xargs` to do the same thing, but i want to minimize
dependency as much as possible. So, using the build-in functionality is
always the priority.

Also, try using `ctrl-/` with `execute()` as a replacement for fzf
preview. The benefit of this is that we can use the entire screen as the
preview window and with our preference fzf height. The downside of
build-in preview window is that it will scale according to fzf height,
so if we setup the fzf height to be small, the preview window is also
gonna be small. Maybe i could use this in vim too? We'll see.