shell(aliasrc): add `less` clear-screen option in `sc` alias
By default, `less` will only do full screen refresh when reaching the
bottom screen. When we haven't reach the bottom screen, `less` won't
do a full screen refresh which can be a problem if we use it in tty.

So let's try adding `clear-screen` option and see how it goes.