shell(gitrc): add `--quiet` and `--index` to `gstp` alias
The benefit of using `--quiet` is that, git will only print out the
conflict message if there's any conflict. So we can be sure if there's
no message at all, it's safe.

The `--index` is for tracking the staged file. So if we already staged
some file and we use `git stash`, with `git stash pop --index`, it will
restore the file with staged status too.