shell(gitrc): use xdg_state_home for all git worktree created
When we put the worktree into the same directory as main worktree,
sometimes there will be a problem. Especially if we use a framework to
create project, like `Nx` in javascript ecosystem, that require strict
directory structure.

So, let's try putting all git worktree into `$XDG_STATE_HOME` and see
how it goes.