vim(mappings): change keybindings put blank line to leader j and k
The problem with previous keybinding, which is `ZN` and `ZH`, is that
sometimes i press the second character first so the cursor moved.

For example, when i want to press `ZN`, sometimes i press `N` first
which lead to trigger normal mode mapping `N` instead, which will move
the cursor to previous search keyword in backward direction.

The decision of using `<leader>j` and `<leader>k` is that `j` represent
down and `k` represent up. So, it should be more intuitive than previous
keybindings.