tmux(conf): use `prefix-b` to create new window with current path
From tmux version 3.5, tmux treat some key as the same key. For example,
`ctrl-m` and `enter` treat as the same keybinding. So, if we want to use
`ctrl-m` as keybinding, we should use `enter` instead of `ctrl-m`
directly.

I want my config to be backward compatible with previous version, and if
i change this keybinding to `enter`, i think it will behave differently
in previous version.

With that in mind, let's try use `prefix-b` and see how it goes.