zsh(zshrc): change default keybinding `alt-d`, `alt-b`, `alt-f`
The default keybinding `alt-d`, `alt-b`, and `alt-f` in zsh depends on
`$WORDCHARS` which is a symbol that represents if the some text
considered as whole word or not. For example, in zsh `anu-itu` is
considered one word because `-` is in `$WORDCHARS`. If we use `alt-d` or
any keybinding that change one word, the whole `anu-itu` will be
effected, not just `anu` or `-itu`. In vim, this is similar to
`iskeyword` option.

This changes is to make `alt-d`, `alt-b`, and `alt-f` to only treat
alphanumeric character as word. For example, if we have `anu-itu`, these
new keybinding will threat `anu` and `-itu` as separate word, instead of
`anu-itu` as whole word. In short, with this changes, `anu-itu` become 2
words instead of 1 word in the default behavior.

If we want to use the default behavior, we can use the upper case
version of it, that is `alt-D`, `alt-B`, and `alt-F`.

This behavior is in line with readline behavior which is used in a lot
of program, such as `bash`.