tmux(conf): add indicator when panes in current window more than 2
We can alternate between current active pane and previously active
pane with `prefix-;`. So, if there are more than 2 panes in current
window, the keybinding `prefix-;` might not be functioning as expected
because we might forgot which pane is the previously active panes.

So, to make switch panes easier, i try adding an indicator in the
status line if the total panes in current window more than 2. If the
indicator is there, i need to use `prefix-q` instead of `prefix-;`.

The indicator is just total panes in current window.