qutebrowser(config): change empty page background to black
When we try to open new website, qutebrowser will open an empty page
with white background which make my eyes hurt.

So, change the empty page background to black. And also, add keybinding
to toggle empty page background to white or black in case changing the
page background to black cause some issue.