i3(config): revert to default color scheme
The problem with the current color scheme is that i use the same color
for background `focused` and `focused_inactive`, and that make me
confused which window is currently focused. The only thing that
differentiate the `focused` and `focused_inactive` is the color of the
text which can be harder if i want to switch between floating and tiling
window.

So, let's try using the default color scheme and see how it goes.