i3(config): add reference brighter grey color for focused window
The distinction between focused and focused_inactive window is much
better because i use brighter grey for focused window and darker grey
for focused inactive window. But the thing is, i still think that the
default color scheme is more practical for day to day use. So i will
still use that.

This is just for in case i want to go back to greyscale color. We'll see
how it goes.