zsh(zshrc): remap `alt-t` to match readline's `alt-t`
`alt-t` in readline is to swap the last 2 words. The definition of word
in readline, as far as i know, is alphabet and numeric. For example,
`anu69` is considered one word.

Because zsh use `$WORDCHARS` env variable to determine symbol that
considered a word, it will also assume anything that has those symbol is
one word. For example, with default `$WORDCHARS`, `anu-69` is considered
one word. Meanwhile in readline, `anu-69` is considered 2 words, `anu`
and `69`. In readline, we can swap the position `anu` and `69` in
`anu-69`.

With that in mind, remap `alt-t` to match readline's `alt-t`. And if we
want the default behavior `alt-t` in zsh, we can use `alt-T` (with
capital T).