tmux(conf): add `previous-prompt` and `next-prompt` keybinding
Replace the previous implementation of previous prompt which use
`search-backward` with regex to new `previous-prompt` command.

The previous implementation has a down side, if the output also start
with `$`, we will go to those output instead. So, using OSC 133 to go to
prompt and the command output is more convenient.

The only down side of using `previous-prompt` and `next-prompt` that i
know is that it's only exist on version 3.4 or later. So if we have an
outdated tmux, we won't be able to use this.

Can we use `search-backward` with OSC 133? I will try this if i have
time, i guess.

Also, add `next-prompt` keybindings too.