vim(pack): add reference vim-shout plugin
`vim-shout` plugin is kind of similar to opening terminal in `acme` text
editor. We can treat the command and output like regular text, which
means we can change the command we run inside of `vim-shout` buffer and
the output of those command.

I'm considering using this plugin but it's only available in vim 9 or
later because it use vim9script, and maybe other features that recently
introduced in vim 9. So for now, i will just put this plugin in the
reference list.