shell(gitrc): change query string when select directory with `gaf`
The purpose of keybinding `ctrl-t` is to make selecting untracked file
inside of directory easier. Sometimes i only want to stage some files in
untracked directory instead of the entire directory.

With this change, we can limit the scope of untracked directory by
using `transform-query()` to change the current query string. If we want
to go back to previous list, we can use `ctrl-r`.