vim(plugin): load netrw on startup
According to this vim commit:
https://github.com/vim/vim/commit/9cfdabb074feefc9848e9f7a4538f201e28c7f06

Seems like `netrw` plugin is moved to `pack/dist/opt/netrw`. So we need
to use `packadd` if we want to use any function in `netrw`, such as
`netrw#BrowseX()` and `netrw#CheckIfRemote()` which is a component for
`gx` keybinding to open url in browser.

Rather than using adding `packadd netrw` in the keybinding, let's try
load `netrw` at startup and see how it goes.