shell(rofibrowser): add script to select browser with rofi
The concept is that when i use `xdg-open` to open url, i want to be able
to select which browser i use to open those url. It might slow down
opening a url a bit, but it can be useful if the browser have problem
with the url or i need credentials to access the url.

For example, sometimes playing video on qutebrowser can be more
demanding than on chrome or other web browser.