i3(config): disable mouse scroll wheel keybindings
Sometimes i place the mouse cursor at the status bar or on the tab of
the window, and when i forgot where the mouse cursor and scrolling, it
will move through workspace (when on status bar) or windows (when on the
tab of the window). So to prevent those unintentional actions, let's try
disable mouse scroll wheel keybindings and see how it goes.