i3(config): set popup_during_fullscreen to leave_fullscreen
Feels like when in fullscreen and if there's a popup window, better off
leaving the fullscreen state and handle those popup window.

Now the question is, how to make the popup window not overlap with other
floating windows? If we use `no_focus` and there's a floating window
bigger than the popup window, we might not be able to see the window
because it's behind the floating window. Need more research for this.