vim(editorconfig): fix nested curly bracket in section pattern
Previously in commit 113a5b3, we already fixed section pattern that has
whitespace in between like this:
```
[{*.txt, *.json}]
```

Here's an example rules of previous problem with whitespace in between:
https://github.com/httpie/cli/blob/master/tests/fixtures/.editorconfig

Now, the problem is, how to handle nested curly bracket section pattern
like this?
```
[{*.{awk,c,dts,dtsi,dtso,h,mk,s,S},Kconfig,Makefile,Makefile.*}]
```

Here's an example rules of current problem with nested curly bracket:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/.editorconfig

Fortunately, the great tpope already handle this on his vim plugin
called `vim-sleuth`. You should check it out on:
https://github.com/tpope/vim-sleuth

So, adjust the mechanism to find the section pattern to follow
`vim-sleuth` plugin. Full credit to tpope.