vim(editorconfig): ignore space in section pattern
Seems like the new pattern does not ignore the space. For example, let's
say we have section pattern like this:
```
[{*.txt, *.json}]
```

It will grep `*.txt` and ` *.json`, the space before `*.json` is also
included which is unexpected.

Not sure if adding a space after comma is a valid syntax or not, need to
confirm the specification. But let's try to ignore it and see how it
goes.

Here's an example of `.editorconfig` file used as reference:
https://github.com/httpie/cli/blob/master/tests/fixtures/.editorconfig