Revert "vim(editorconfig): ignore space in section pattern"
This reverts commit 9483a8f4649bf756bf09bbdf5e16b360e52da593.

If we use reference from curly bracket on `bash`, `zsh`, or other shell,
it seems like there should be no space after the comma.

For example, in `bash`, this will print `anu itu`:
```bash
echo {anu,itu}
```

But this will print `{anu, itu}:`
```bash
echo {anu, itu}
```

Basically if there's a space inside the curly bracket, `bash` will print
it "as is" rather than doing an expansion.

With that in mind, let's revert this changes for now and assume the
previous example is invalid.