vim(mappings): change alternate buffer key binding to ctrl-m
The problem with `ctrl-s` is that some terminal use it as flow control,
like in tty, which means that we can't use it in those terminal. And it
turns out i don't really use default key binding `ctrl-m` or `enter`
that much. So let's try using `ctrl-m` as alternate buffer key binding
in normal mode and see how it goes.