vim(mappings): use <leader>space for alternate buffer key binding
Before this i use `ctrl-m` as a replacement for previous alternate
buffer `ctrl-s` which i can't use in tty. Now, the problem with `ctrl-m`
is that i need 2 hands to press the key, or if i use `enter` key (which
is equivalent to `ctrl-m` in some terminal), i would need to leave the
home row, which is to slow for something i use often.

So let's try using `<leader>space` for alternate buffer normal mode
key binding and see how it goes.