shell(gitrc): make fzf cursor go down after delete branch in gco
The problem is that when we refresh the list, the cursor is in the same
position and the item at the top is going down. That make me
accidentally delete the git branch that i don't want to delete.

So let's try moving the cursor down after refresh the git branch list
and see how it goes.