shell(gitrc): change git stash list format in gstl function
The problem with `--oneline` format is that the stash name have trailing
`:` (colon) in it, so it looks like this:
```
commit-hash refs/stash@{0}: ...
```

And that's why i use the commit hash which is on the first column for
preview the content of the stash. Sometimes the preview does not work
using those commit hash, but it works if i use `refs/stash@{0}` or
`stash@{0}`.

So let's try change the git stash list format like this:
```
stash@{0} ...
```
so we can use the reference `stash@{0}` on the first column directly
and see how it goes.