shell(gitrc): add ctrl-s to pop git stash from gstl function
So that we don't need to copy the stash reference and manually pop the
stash using `git stash pop` command or `gstp` alias.

I use `become()` in fzf so that we're not staying in fzf after we run
`git stash pop` command. This is to prevent we get git conflict when
applying stash while inside fzf, which fzf will ignore and make us not
realize we have a conflict.