vim(mappings): give feedback if keyword to grep is not found
The problem with the previous grepping mechanism is that there's no
feedback if the keyword we're trying to search is not found, which means
that there's no indicator if the grepping is still on going or already
finished but there's no match for the provided keyword.

So, add a message to indicate that there's no match for the provided
keyword if there's no result, This is also to distinguish no result from
long running grep process in large code base.