tmux(conf): unbind space keybinding
The default space keybinding is used to change layout to the next
predefined layout, and because i use `ctrl-space` as prefix, often times
i accidentally hit space and change the current layout.

So, to prevent unintentional layout change, let's try unbind the default
space keybinding and see how it goes.