i3(config): use variable for focus direction
This is inspired by `sway` config, which is still in progress.

Previously my home row on the right side is on `hjkl`, but now i want to
move it to a proper home row `jkl;`. Currently i'm still training to
type with the proper home row, but once i get used to it, i might change
the focus direction using `jkl;` instead of `hjkl`. That's why i use
a variable to make changing focus direction easier.