vim(mappings): xmap to use visual selection as substitution input
Sometimes the words i want to substitute is in the middle of a line, so
to make things easier to select, i use visual mode. The problem is that,
we need to yank and put the selected word into substitute command. I
don't really want to replace the `0` register if possible.

That's why let's try adding visual mode keybinding to directly put the
words from visual selection into the substitute command without
replacing `0` register and see how it goes.