i3(config): put the status bar at the bottom
I plan to put my webcam while streaming on the bottom, so let's try
putting the i3 status bar at the bottom so that the window getting
raised up a little bit and hopefully the webcam didn't block too much
content on my screen. Let's see how it goes.

Also, because we're moving the status bar at the bottom, we also need to
adjust the position of `dunst` (for notification).