tmux(conf): put the status bar in the bottom
This is also for live stream. To prevent my webcam, which i plan to put
at the bottom, blocking more screen content, let's try putting the
status bar in the bottom.

This is also to make the terminal view raised up a bit, which might help
in viewing the command output or opening vim.

The problem that might arise from this change is that, now we have a lot
of status bars in the bottom which might be confusing for some people.
So, let's see how it goes.