vim(fugitive_log): add --follow when select current file
So that we can have the log before renames too.

For example, let's say we have file `anu`, and then we rename that file
to `itu`. Without `--follow`, we only get the log until we rename file
`anu` to `itu`. We didn't get the log when the file still named `anu`.
With `--follow`, we also get the log when the file still named `anu`.

`--follow` only works on one file, hence we can only at this flag to
current file option.