qutebrowser(config): add navigate up and strip keybindings
`navigate up` is to go up a level in the current URL. For example, let's
say we open a url `https://github.com/bruhtus/dotfiles`, when we use
`:navigate up`, we will go to `https://github.com/bruhtus`.

`navigate strip` is to strip query params and fragment from the current
url. For example, let's say we open a url
`https://github.com/bruhtus/dotfiles#resources`, when we use `:navigate
strip`, we will go to `https://github.com/bruhtus/dotfiles`.

For now, the keybinding for `navigate up` is `gu` and the keybinding for
`navigate strip` is `gh`. Let's see how it goes.