shell(aliasrc): add campv() to show webcam footage with mpv
This is for streaming. Rather than show webcam in OBS, let's try show
webcam in mpv.

The benefit of using mpv instead of OBS is that, we can put the mpv
window in i3wm scratchpad to hide the webcam temporarily and make the
mpv window fullscreen with the usual keyboard shortcut.

Let's see how it goes.