i3(config): use shell script to add pomodoro in i3status
The benefit of putting pomodoro in `i3status` is that we can see the
status of pomodoro without opening other program.

Previously i don't want to use this because using script for custom
`i3status` make the process increase according to the display. For
example, when we only use `i3status` the system process registered is
only 1 even when we use 2 monitors. If we use script as a custom
`i3status`, the system process become 2 if we use 2 monitors. The CPU
and memory usage seems negligible so let's see how it goes.

Also because we are using script, adding color become more complex and i
don't really need color. So, the script does not have parsing color
support.