shell(gitrc): remove git commit alias
Turns out that there's a command called `gc`, which is a garbage
collector for C and C++. To prevent conflict with that package, let's
just remove the `gc` alias for `git commit` and use `git commit`
directly.

While we're at it, let's also remove `gca` alias for
`git commit --amend` and use the full command instead. Let's see how it
goes.