i3(config): replace pomo with simo for pomodoro timer
Recently i made my own pomodoro timer in golang for `i3status` and
possibly any other status bar. So replace `pomo`, which the previous
pomodoro timer in bash script, with `simo` which is the new pomodoro
timer in golang.

The problem with `pomo` is that there's no dedicated subcommand for
specific `focus` or `break` session, and because it's bash script,
it's mostly depends on gnu core utils, which might not be available.

With that in mind, let's try using the new pomodoro timer, which is
called `simo`, and see how it goes.