vim(ftplugin): set c filetype indent to linux kernel coding style
From the docs:
https://www.kernel.org/doc/html/v4.10/process/coding-style.html

The default indentation is 8, which is the same value as default
tabstop. So we're using 8 characters to indent.

Setting `sts` or `softtabstop` is so that we can use tab to insert some
space instead of tab character. Looking at the linux kernel code base,
i found that some people also using space after indenting using tab,
just to make align it with the previous line column. That's why
`softtabstop` is set to 2, so that when i press tab, it will insert 2
space when i need to add extra space. To insert tab character, we can
use `ctrl-t` and `ctrl-d` in insert mode or `>>` and `<<` in normal
mode.

Let's see how it goes.