vim(indent): add license in markdown.vim
Seems like there's a change in maintainer ship of the github repo, so
update the reference and add license.

Also, use `endif` instead of `end` to make it clearer which end section
it is, even though `end` is also a valid syntax to end the if condition.
For more info on `:h :end`.