vim(mappings): add searching and found message in custom_grep()
The "searching..." message is to indicate that the searching still on
going and the "found keyword..." message to indicate if the searching
already done and found the keyword we are looking for.

We put the message in message history (more info :h message-history) so
that we can check the previous searched keyword with command
`:messages`.