i3(config): add settings and keybinding for simo-gui
Recently i created `simo-gui` to show the current simo state for live
streaming. So this changes add some settings for `simo-gui` window and
keybindings to move `simo-gui` and `campv` into scratchpad at the same
time.

We need to move the `simo-gui` and `campv` into scratchpad at the same
time so that we can move out those 2 windows from scratchpad as a
bundle.

Also use `no_focus` with `simo-gui` to prevent focusing on the window
when running `simo-gui`.