Revert "vim(lsp): adjust toggle lsp with new vim-lsp-settings init mechanism"
This reverts commit 9c570aaf82543ca0b801fb11b1d9289ea3218987.

Seems like in commit 928fe58, vim-lsp-settings using the old init
mechanism as the default init and if we want to use the new lazy
loading mechanism, we need to add global variable like this:
`let g:lsp_settings_lazyload = 1`.

With that in mind, revert to old setup for vim-lsp-settings because i
have some concern when using `doautocmd` on all `BufNewFile` and
`BufEnter` autocmd events. Let's see how it goes.