shell(sdfm): change sdfl alias to function for zsh files completion
Currently vim-fugitive does not support `--git-dir` and `--work-tree`
flag for git, so the roundabout way to get the log of specific file is
using `sdfm log` with the filename.

To make that happen, i need to assign the completion `_files` to `sdfl`
in zsh. But the thing is, i only know how to assign the completion for
the shell function, i am not sure if we can assign the completion for
shell alias. So change the `sdfl` shell alias to shell function and
assign completion `_files` to `sdfl` shell function.

Let's see how it goes.